agrega lógica y permisos de usuarios premium y admin.

+++ src/dao/controllers/product_controller.js
@@ -82,9 +82,14 @@ export const updateProduct = async (req, res) => {
 }
 
 export const deleteProduct = async (req, res) => {
-    await producto.deleteProduct(req.params.pid);
-    req.logger.info(`products deleted - ${new Date().toLocaleTimeString()}`);
-    res.status(200).json({ message: 'Producto eliminado' });
+    const currentProduct = await producto.getProductById(req.params.pid);
+    if (currentProduct.owner === req.session.user) {
+        await producto.deleteProduct(req.params.pid);
+        req.logger.info(`products deleted - ${new Date().toLocaleTimeString()}`);
+        res.status(200).json({ message: 'Producto eliminado' });
+    } else {
+        res.status(401).json({ error: 'Usuario no autorizado' });
+    }
 }
 
 export const addProduct = async (req, res) => {
